fix: update authorized cta button styling, refactor into separate component
