+++ src/Scenes/Account/Account.tsx
@@ -1,6 +1,6 @@
-import { Box, Button, Container, Flex, GuestView, Sans, Separator, Skeleton, Spacer } from "App/Components"
-import { Schema as NavigationSchema } from "App/Navigation"
+import { Box, Container, Flex, GuestView, Sans, Separator, Skeleton, Spacer } from "App/Components"
 import { useAuthContext } from "App/Navigation/AuthContext"
+import { Schema as NavigationSchema } from "App/Navigation"
 import { Schema, screenTrack, useTracking } from "App/utils/track"
 import { ChevronIcon } from "Assets/icons"
 import {
@@ -17,22 +17,17 @@ import {
   Star,
 } from "Assets/svgs"
 import gql from "graphql-tag"
-import { DateTime, Duration } from "luxon"
+import { DateTime } from "luxon"
 import { default as React, useEffect } from "react"
 import { useQuery } from "react-apollo"
 import { Linking, Platform, ScrollView, StatusBar } from "react-native"
 import * as Animatable from "react-native-animatable"
 import Share from "react-native-share"
-import styled from "styled-components/native"
 import { State, UserState } from "../CreateAccount/CreateAccount"
-import { HourMinuteSecondCountdown } from "./Components/HourMinuteSecondCountdown"
 import { InvitedFriendsRow } from "./Components/InviteFriendsRow"
 import { NotificationToggle } from "./Components/NotificationToggle"
+import { AuthorizedCTA } from "./Components/AuthorizedCTA"
 import { AccountList, CustomerStatus, OnboardingChecklist } from "./Lists"
-
-const SansUnderline = styled(Sans)`
-  text-decoration: underline;
-`
 
 export const GET_USER = gql`
   query GetUser {
@@ -86,10 +81,10 @@ export const GET_USER = gql`
 `
 
 export const Account = screenTrack()(({ navigation }) => {
-  const tracking = useTracking()
-
   const { authState, signOut } = useAuthContext()
   const { data, refetch } = useQuery(GET_USER)
+
+  const tracking = useTracking()
 
   useEffect(() => {
     const unsubscribe = navigation?.addListener("focus", () => {
@@ -109,7 +104,6 @@ export const Account = screenTrack()(({ navigation }) => {
   const onboardingSteps = customer?.onboardingSteps
   const status = customer?.status
   const referralLink = customer?.referralLink
-  console.log(referralLink)
 
   const user = customer?.user
   const email = user?.email
@@ -282,77 +276,35 @@ export const Account = screenTrack()(({ navigation }) => {
           />
         )
       case CustomerStatus.Authorized:
-        const authorizedAt = DateTime.fromISO(customer?.authorizedAt)
-        const authorizationWindowClosesAt = DateTime.fromISO(customer?.admissions?.authorizationWindowClosesAt)
-        const targetAuthorizationDate = authorizationWindowClosesAt.isValid
-          ? authorizationWindowClosesAt
-          : authorizedAt.plus({ days: 2 })
-        const authorizationDuration =
-          targetAuthorizationDate.valueOf() > authorizedAt.valueOf()
-            ? targetAuthorizationDate.diff(authorizedAt, "hours")
-            : Duration.fromMillis(0)
-
         return (
-          <Box pb={1}>
-            <Flex alignItems="center" pb={3}>
-              <HourMinuteSecondCountdown targetDate={targetAuthorizationDate} />
-            </Flex>
-            <Sans size="2" color="black100" textAlign="center">
-              You're in. Let's choose your plan
-            </Sans>
-            <Spacer mb={1} />
-            <Sans size="1" color="black50" textAlign="center">
-              You have{" "}
-              <SansUnderline size="1" color="black50">
-                {authorizationDuration.get("hours") === 1
-                  ? `${authorizationDuration.toFormat("h")} hour`
-                  : `${authorizationDuration.toFormat("h")} hours`}
-              </SansUnderline>{" "}
-              to secure your spot. If we don't hear from you, your invite will go to the next person and{" "}
-              <SansUnderline size="1" color="black50">
-                you'll be waitlisted
-              </SansUnderline>
-              .
-            </Sans>
-            <Spacer mb={3} />
-            <Flex flexDirection="row" justifyContent="center">
-              <Button
-                variant="primaryWhite"
-                onPress={() => {
-                  tracking.trackEvent({
-                    actionName: Schema.ActionNames.LearnMoreTapped,
-                    actionType: Schema.ActionTypes.Tap,
-                  })
-                  navigation.navigate("Modal", {
-                    screen: NavigationSchema.PageNames.CreateAccountModal,
-                    params: {
-                      initialState: State.ChoosePlan,
-                      initialUserState: UserState.Admitted,
-                      onMountScrollToFaqSection: true,
-                    },
-                  })
-                }}
-              >
-                Learn more
-              </Button>
-              <Spacer mr={1} />
-              <Button
-                variant="primaryBlack"
-                onPress={() => {
-                  tracking.trackEvent({
-                    actionName: Schema.ActionNames.ChoosePlanTapped,
-                    actionType: Schema.ActionTypes.Tap,
-                  })
-                  navigation.navigate("Modal", {
-                    screen: NavigationSchema.PageNames.CreateAccountModal,
-                    params: { initialState: State.ChoosePlan, initialUserState: UserState.Admitted },
-                  })
-                }}
-              >
-                Choose plan
-              </Button>
-            </Flex>
-          </Box>
+          <AuthorizedCTA
+            authorizedAt={DateTime.fromISO(customer?.authorizedAt)}
+            authorizationWindowClosesAt={DateTime.fromISO(customer?.admissions?.authorizationWindowClosesAt)}
+            onPressLearnMore={() => {
+              tracking.trackEvent({
+                actionName: Schema.ActionNames.LearnMoreTapped,
+                actionType: Schema.ActionTypes.Tap,
+              })
+              navigation.navigate("Modal", {
+                screen: NavigationSchema.PageNames.CreateAccountModal,
+                params: {
+                  initialState: State.ChoosePlan,
+                  initialUserState: UserState.Admitted,
+                  onMountScrollToFaqSection: true,
+                },
+              })
+            }}
+            onPressChoosePlan={() => {
+              tracking.trackEvent({
+                actionName: Schema.ActionNames.ChoosePlanTapped,
+                actionType: Schema.ActionTypes.Tap,
+              })
+              navigation.navigate("Modal", {
+                screen: NavigationSchema.PageNames.CreateAccountModal,
+                params: { initialState: State.ChoosePlan, initialUserState: UserState.Admitted },
+              })
+            }}
+          />
         )
       case CustomerStatus.Invited:
       case CustomerStatus.Active:
